requirements.txt
selenium==4.5.0
pytest==7.1.2
pytest-xdist==2.5.0
allure-pytest==2.9.45